Added configuration
Necessary to provide all attributes for configuring a logger upon initialization.

The Reek and RuboCop issues are a concern but are not warranted for this object since the problems stem from the original design of the `Logger` class and is being indirectly surfaced in this design. I will note that `ModuleInitialize` (Reek) is a bug and should be fixed by Reek (I have an open issue on this subject).