Added sanitizer utility
Necessary to transform a positional log entry into a keyword log entry for further use by formatters and their associated templates. By default, these log entries consist of: severity, at, id, and message. In order to support any number of attributes -- especially when the message is a hash payload -- the message needs to be splatted into the transformed hash for further processing.

It's important to note that the `severity`, `at`, and `id` keys are added to the message hash in order to overwrite any idential keys since these keys are protected and carry critical log entry metadata.