Added registry
Necessary to provide a global registry for the global configuration of this gem.

The Reek and RuboCop issues are related to the `.extended` method which is definitely longer than I'd like but provides a nice way to load all defaults via a single method which doesn't make sense to break up further.