Fixed sanitizer to always include message
Ensures situations where templates are expecting the message key/value pair, but never get it, don't raise an exception. The message can still be overwritten, if provided, but will be `nil` in these cases which won't cause an error with `Kernel#format`. Example:

    format "%<message>s"
    # one hash required (ArgumentError)

    format "%<message>s", message: nil
    # ""

This reduces the number of log crashes for these situations and improves the experience.