Updated formatters to consume log entry
Necessary to leverage the new log entry object. An instance of `Cogger::Entry` is now passed to all formatters for processing. Each entry includes the program ID, severity, and date/time which is why none of this information needs to be _explicitly_ passed to the formatters. If explicit program ID, severity, and date/time is passed to a formatter, it's ignored since the entry is now the source of truth. The rest of the related changes are highlighted below:

- The hub accepts a message and payload. The payload can be any arbitrary key/value pair for additional logging information. Reek was updated to ignore the data clump for all of the log severity helper methods since it can't be avoided.
- The registry has been updated to account for `ANY` severity instead of `UNKNOWN` which is now compatible with all formatters as well as the original Ruby logger.
- The sanitizer has been simplified since each log entry can be filtered by payload and each entry can be asked for it's corresponding attributes (i.e. filtered payload).

Milestone: minor