Added tagged logging
Necessary to allow tags to be added to logs, if desired, which builds upon the previously added `Tag` and `Entry` objects. Since tags can be combined (i.e. when the logger is initialized and when writing a log), it is up to each formatter to process tags as desired. I've added tags to all formatters tha can leverage them.

Milestone: minor