Added Rack logger
Necessary to enhance default logging with information plucked from the current request based on custom tags. The logger is updated, accordingly, once the tags are resolved for the current request.

Rack is _implicitly_ required in order to not add another gem dependency to this gem while still making this implementation available to Rack-based web frameworks that might need request logging.

Dependency injection could be cleaned up if link:https://alchemists.io/projects/infusible[Infusible] was added as a dependency but am holding off, for now, in favor of keeping dependencies low even though use of a primitive hash for the dependnecy container is not ideal.

Milestone: minor