Added time unit
Necessary to calculate the unit of time for a given duration. This will soon be used by a timer to calculate duration along with unit.

Use of RuboCop Style/MethodCallWithArgsParentheses has been temporarily disabled since this is a RuboCop bug that I've filed with the team to look into further.

Milestone: minor