Added hub abort
Provides a convenient solution to a common Command Line Interface (CLI) pattern where you want to log an error message _and_ exit the program at the same time so the terminal status code matches the severity of the log message.

The specs use `SystemExit` in order to not cause the entire test suite to exit. This could be handled differently by injecting `Kernel` into the logger configuration but opted to avoid adding another configuration attribute by rescuing `SystemExit` instead.

The Reek issue with too many methods is a concern. For now, leaving this as a reminder to think on and resolve better in the future.

Milestone: minor