Fixed configuration to freeze tags upon initialization
Necessary to deal with a discrepancy when you create a configuration with default (empty) tags, you can't append to them because they are frozen by default. Alternatively, if you supply custom tags, you _could_ append to them which is not desired due to confusing behavior especially if you are using the same logger instance across multiple objects. So tags are always frozen now.

Milestone: patch