Added property formatter
Necessary to support the _informal_ link:https://brandur.org/logfmt[logfmt] as used by Heroku and other services. The redundant use of "log" and "fmt" (short for format) in "logfmt" is unnecessary within the domain of a logger so this formatter is called `Propery` since use of `key=value` syntax is more akin to _property_ or _INI_ style syntax found in UNIX like environments or Java property files.

Milestone: minor