Fixed registry to add formatters with fully resolved templates
Previously, when asking the registry for formatters, we only got the provided formatter and associated template as added to the registry upon load. In most cases, the template was `nil` which wasn't accurate because most formatters have a _default_ template.

This change ensures the registry properly represents the formatter _and_ associated template by falling back to the default template as supplied by the formatter when a custom template isn't registered with the formatter. This has the added benefit of making the `Registry#formatters` more accurate as well.

In situations where a formatter is registered but doesn't providate a default template via the `TEMPLATE` constant, a `NameError` will be raised with clear instructions on how to fix formatter's implementation.

Reek was updated to ignore the control parameter because, in this case, falling back to the formatter's default template (constant) is what we want.

Milestone: patch