Added key format transformer
Necessary for transforming the target into a key that can be used for looking up colors or emojis.

This is an evolution of the original `Formatters::Kit::Colorizer` but specialized for use within this new `Transformers` namespace. The original version will soon be deleted once all of the new transformers are implemented.

Milestone: minor