Fixed entry message to use question marks for unknown characters
Necessary to ensure we don't end up with exceptions or fatal crashes in the logs when non-UTF-8 characters show up in the log entry messages. This situation is most common when dealing with binary downloads such as file attachments, images, etc. especially when using the link:https://github.com/httprb/http[HTTP] gem.

To make this possible, the private `.sanitize!` method is introduced which ensures a message uses `UTF-8` encoding where all unknown characters are replaced with questions marks. This method also makes it possible to use a `Hash` for the message which will be extracted and properly encoded while allowing all additional attributes to be merged into the payload. This is most beneficial when using the JSON formatter.

Milestone: patch